staging.py: enable pushing the branch and tag with --ssh-key
Add last step to force-push the resulting branch with newly created
tag to the origin.  Also add an --ssh-key option to specify an SSH key
to use for this purpose, which is mainly useful for automation.  If
this option is not specified, the regular user keys and ssh
configuration will be used as expected.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>