ansible: add script to update the staging backend or frontend
Add "ansible" script to make it easy to update kernelci-backend or
kernelci-frontend.  If an SSH key is present in
keys/id_rsa_staging.kernelci.org then it will use it automatically.

To run this with all default values, to update staging instances with
the staging.kernelci.org branch as created by pending.py:

  ./ansible kernelci-backend
  ./ansible kernelci-frontend

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>