kernel.py: add script to create a kernel test branch
Add kernel.py to create a kernel test branch with patches applied on
top of a remote branch, and push it with a tag to the kernelci Github
linux repository.  This is meant to be used for branches to test on
staging.kernelci.org and for testing production updates on
kernelci.org.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>