patches/kernelci-core: fix kselftest in kernelci_staging
Fix the kernelci_staging config to enable the kselftest fragment.

Fragments that are generated such as kselftest.config need to be put
in the source tarball, and therefore can't be used only on a
per-architecture basis at the moment.  So they need to be enabled for
the whole build configuration variant.  Doing this extends the number
of builds on staging so it's only a temporary measure until fragments
can be set with more granularity.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>