kernelci.date_tag: fetch tags from origin before creating new one
Fetch the tags from the origin where the new branch will be pushed to,
rather than when checking out the initial repository.  This solves the
case of the kernel branches that are based on an upstream branch but
the created tag is pushed to the kernelci repository.  The tags need
to be fetched from the kernelci repository, not the the upstream.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>