staging.kernelci.org: add cmd_rotate_tree()
Add a rotate_tree command to rotate the kernel tree to use between
mainline, stable and next and use it in cmd_all() to rotate by
default.

Also ignore the .tree file which is used to locally keep the name of
the last tree used, to rotatate automatically every time the script is
called.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>