job.py: add abort command
Add "abort" command to abort all the running builds of a given job
name. It will ask Jenkins to abort them but not wait for them to stop,
so in some cases it may be needed to run this several times.  It's
only to be used as a convenience for manual intervention.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>