job.py: add flush_kernel_trigger_jobs command
Add the flush_kernel_trigger_jobs command to job.py to cancel jobs
that have no artifacts, which is the case with kernel-build-trigger
jobs that have not started running yet or have not yet retrieved lab
info.  Such jobs can be cancelled as the next monitor job will restart
them.

This helps with production updates when currently running trigger jobs
need to complete, but others that haven't started can wait until after
the update is complete (and the monitor is disabled during that time).
There is no simple Jenkins API to get the list of jobs that are
waiting for an executor, since they are already in the "running" state
once scheduled.  So this hack is necessary here.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>