patches/kernelci-core: add fix with separate chromeos-configs
Add a temporary patch to split the chromeos-configs into a separate
entry in the YAML build configuration so it can be used with any
branch, rather than only ones defined after in the file.  This should
also make it easier to maintain, and it should eventually get merged
into the chromeos branch.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>