Fix buildroot patches path
It's required for the patches that are applied to a specific branch to
be in a directory named after that branch. A patch that needs to be
applied for staging branch of buildroot needs to be moved to
staging.kernelci.org directory. Otherwise pending.py script will not
apply it.

Signed-off-by: Michal Galka <michal.galka@collabora.com>