kernelci.apply_patches(): also apply .mbx files and recursively
The b4 utility provides a way to download a patch series as an .mbx
file which can be directly applied with "git am".  Replace the glob
command which looked for only *.patch files with a pattern matching
both *.patch and *.mbx including sub-directories.  This will also
allow putting series of individual patch files in their own
directories to make things easier to manage.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>