kernelci.org: automatically update rootfs URLs
Trigger rootfs build jobs with a pipeline version and use that to
update the URLs automatically in test-configs.yaml once the jobs have
completed.  Create a branch and push it to GitHub if the new URLs are
valid.

This doesn't create a PR automatically yet, doing so would require
another Python command using the GitHub API.  This can be done as a
follow-up.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>