chromeos.kernelci.org: build cros-* Docker images
Update the list of Docker images to build with their new names
starting with the cros- prefix.

The chromeos-flash image hasn't been renamed yet as it depends on a
rootfs config to be merged in the main branch first with a public SSH
key matching the private one in the Docker image.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>