patches/kernelci-pipeline: create staging config file
Replace all the patches in config/kernelci.conf with a single one
which adds config/staging.kernelci.org.conf with staging specific
configuration.  This avoids unnecessary conflicts when changes happen
in the default config file, which isn't really meant to be used as-is
anyway except as a reference for a local development instance.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>