staging.kernelci.org: drop kernelci fragment with dt-validation
The dt-validation requirements include PyYAML as well as the
kernelci-core Python package.  As a result, when building
dt-validation image with the kernelci fragment enabled there is a
conflict and pip3 fails to uninstall the current PyYAML.

Avoid this issue for now by building dt-validation without the
kernelci tools.  Ultimately this should be fine when running in
Kubernetes as separate steps anyway.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>