staging.kernelci.org: only restart API if there were changes
Create a temporary local integration branch to look for a diff with
the current staging branch, then only restart the API if there are
changes.

That way, the API services don't need to be restarted at every staging
job and client scripts running elsewhere can remain connected.  They
will need to reopen their connections if the API is restarted, but
that's typically because of API changes so client-side changes are
most likely required too anyway.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>