patches/kernelci-pipeline: adjust trigger settings
Adjust settings for the trigger service to not use --force to avoid
testing the same revision twice, drop the build_config as now the
service iterates over all the configs and set a polling period of 1h
to actually monitor new revisions.  Typically, this will just detect
some mainline updates as the staging ones are always updated just
before the pipeline is restarted.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>