staging.kernelci.org: use master branches for test definitions
The test-definitions and iec-security upstream repositories haven't
moved from master to main branches so align the forks to also user
master rather than main.  This is now the convention for forks to use
the same branch name as the upstream and that's the branch that will
get automatically rebased in the fork during each production update.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>