staging.kernelci.org: Build architecure specific clang containers
As part of enabling coverage of kselftest for clang builds
archiecture specific Docker containers are being added containing
libraries needed to build real userspace applications.  Include
these in the set of containers we build for staging, leaving the
existing containers in place to avoid disrupting existing usage.

Similar updates will be needed to the production script to deploy
this but to minimise the risk of disruption there those will be
done separately if this approach progresses.

Signed-off-by: Mark Brown <broonie@kernel.org>