kernelci.org: drop kci_docker --verbose argument
The --verbose argument when running kci_docker is mostly useful on
staging when things are likely to break often.  For production builds,
it's actually better to have a clear log output with the list of
images being built without all the noise.  Also it's easy enough to
restart an image build by hand with --verbose if something goes wrong,
to investigate.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>