__init__.py: Enable optional Github token support
This commit adds functionality to the code to check for the presence
of a GitHub access token in the keys/gh-token.txt file.
If the token is present, it will be used to raise the rate limits
for GitHub API requests. If the token is not present,
the code will continue to behave as before without raising the rate limits.

This change should help to avoid hitting the rate
limits when making multiple API requests to GitHub.

Signed-off-by: Denys Fedoryshchenko <denys.f@collabora.com>