patches/kernelci-core: trim down the number of staging builds
Rework build-configs-staging.yaml to remove all the builds that aren't
really useful on a daily basis for KernelCI development and reduce the
burden on the infrastructure.  Now we have 6 for stable, 13 for
mainline and 20 for linux-next.  This also adds one rustc build which
was not being covered at all and is actually useful while rustc
support in the kernel and enabling coverage in KernelCI is still being
worked on.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>