staging.kernelci.org: use settings file with secrets
Runtime environments as well as the API require secrets such as
tokens.  The standard way of providing them is via the settings file,
but until now this was done solely via environment variables in
kernelci-pipeline on staging.  We're hitting the limits of this
approach with LAVA labs as each of them requires a different API
token.

To avoid this issue, switch to using a file not under source control
which can contain secrets such as LAVA API tokens.  We might then
start adding it as an encrypted file for a more longer-term solution.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>