kernelci.org: add rootfs_docker command
Build the Docker images used for building rootfs images just before
building the rootfs images rather than as part of the whole Docker
image build step.  This is required to use the main branch and update
the rootfs images with the latest changes ahead of the rest of the
production update which relies on test results with the new rootfs
images to be deployed.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>