staging.kernelci.org: build KernelCI API image
This change depends on kernelci/kernelci-core#2332

Updating API branch has been moved to the beginning of the run. It is
meant to remove the possibility of pushing other images to the Docker
registry than will be deployed on the instance. This will become
irrelevant once deployment is migrated to versioned images instead of
docker-compose-built ones.

Signed-off-by: Paweł Wieczorek <pawiecz@collabora.com>